fixed several "conversion from 'lua_Integer' to 'int', possible loss of data" warnings

+++ plugins/config/holdes_mondenkind.pyx
@@ -1692,7 +1692,7 @@ cdef int handle_cmd_args(lua_State *L, list parameters):
     return True
 
 
-cdef int handle_cmd_s_o_e_code(lua_State *L, int error_code, dict error_codes):
+cdef int handle_cmd_s_o_e_code(lua_State *L, Py_ssize_t error_code, dict error_codes):
     """
     {
         id = "ERROR_SUCCESS",
@@ -1731,7 +1731,7 @@ cdef int handle_cmd_s_o_e_codes(lua_State *L, dict error_codes):
     }
     """
     cdef:
-        int error_code
+        Py_ssize_t error_code
         unicode u_description = u""
     # first key
     lua_pushnil(L)
